backend/requirements.txt
# Core ML
torch==2.0.1
torchvision==0.15.2
timm==0.9.12

# Image Processing
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
albumentations==1.3.1
Pillow==10.1.0

# Web Framework
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1

# Data Processing
numpy==1.24.3
pandas==2.0.3
scikit-learn

# Utilities
tqdm==4.66.1
requests==2.31.0
python-dotenv==1.0.0
huggingface-hub==0.19.4

# Production Server (optional)
gunicorn==21.2.0

# Background removal / inference
rembg==2.0.50
onnxruntime==1.16.3

# Grooming recommendation (Gemini Vision AI)
google-generativeai==0.8.0
